requirements.txt
# ── PyTorch（CUDA 11.7）──────────────────────────────────────────────────────
# 需通过 PyTorch 官方源安装，pip install 无法直接解析 +cu117 后缀：
#   pip install torch==1.13.1+cu117 torchvision==0.14.1+cu117 \
#       --index-url https://download.pytorch.org/whl/cu117
# torch==1.13.1+cu117
# torchvision==0.14.1+cu117

# ── 图像处理 ─────────────────────────────────────────────────────────────────
numpy==1.21.6
opencv-python==4.13.0
Pillow==9.2.0
albumentations==1.3.1

# ── 训练工具 ─────────────────────────────────────────────────────────────────
tqdm==4.45.0

# ── 超参数调优 ───────────────────────────────────────────────────────────────
optuna==4.0.0
optuna-dashboard==0.16.2     # 可选，实时监控 Optuna 调优过程

# ── 实验监控 ─────────────────────────────────────────────────────────────────
wandb==0.18.7                # 可选，config.yaml 中 wandb.enabled=false 时不需要

# ── 配置文件 ──────────────────────────────────────────────────────────────────
PyYAML==6.0.1
